Added missing th.original cell to match td.original.
This fixes the table body (tbody) to be shifted by one column to the
left.  While the width of td.original and th.original is set to 0 in the
CSS, th.original was not defined.

It removes the arbitrary(?) addition of `colspan=2` to the first td,
which gets done only if it is not hidden.

The issue gets triggered by using
[django-concurrency](https://github.com/saxix/django-concurrency/),
which adds a hidden input, that comes first in the tabular inline.