Fixed #25637 -- URLValidator hostname length validation
URLValidator now validates the maximum length of a hostname in total
and the maximum length of all labels inside the hostname. The label
validation is done in the regular expression while the hostname length
check is implemented by introducing a new check in the validator's
__call__ method. Please note that the additional call on urlsplit()
introduced there should not be a measurable performance overhead, as
urlsplit is called anyway in some of the code branches above and the
standard-library implementation uses an internal result cache.