Fixed #21221 -- Made forms Media use staticfiles to serve files if available.
Previously form assets defined as a forms media have been served using a joined
absolute path consisting of the `STATIC_URL` and defined relative path.
This commit changed the `absolute_path` method to serve form media using
`django.contrib.staticfiles` if configured.

This will allow users to server their form media from a CDN or any other
staticfiles storage of their choosing.

For the purpose of #21221 a optional depenedency had to be
made from core to contrib.statifiles. Hence static was
implemented identical to the one in `staticfiles` or `admin`.
Because of that, the following changes have been made:

- Moved admin_static implementation to core incl. deprecation
- Moved staticfiles patch into StaticNode
- Deprecated StaticFilesNode
- Deprecated staticfiles static template tag
- Removed prefix argument from django.forms.widgets.Media.absolute_path
- Changed static template tag documentation

Fixed #18455