Fixed #25761 -- Re-raised exceptions need __cause__.__traceback__
When Django re-raises an exception, it sets the __cause__ attribute even
in Python 2, thus mimicking Python's 3 behaviour for "raise Foo from Bar".
However Python 3 also ensures that all exceptions have a __traceback__
attribute and thus the "traceback2" Python 2 module (backport of Python
3's "traceback" module) relies on the fact that whenever you have a
__cause__ attribute, the recorded exception also has a __traceback__
attribute.

This is breaking at least testtools which is using traceback2 (see
https://github.com/testing-cabal/testtools/issues/162).

This commit fixes this inconsistency by ensuring that Django sets
the __traceback__ attribute on any exception stored in a __cause__
attribute of a re-raised exception.