Fixed #23372 -- made loaddata faster in cases no fixtures match
Django's test suite often tries to load fixture files from apps that have
no fixtures at all. This leads to a lot of non-necessary disabling and
enabling of constraints which can be expensive on some database.

To speed up the above mentioned case, loaddata now first checks if any
fixture file matches. If no fixture file is matched, then the command
exits before disabling and enabling of constraints is done.

The main benefit of this change is seen on MSSQL, where tests on 1.8
run about a order of magnitude faster after this change.