Changed `action="."` to `action=""` in tests and docs.
`action="."` strips query parameters from the URL, which is not usually what
you want.  Copy-paste coding of these examples could lead to difficult to
track down bugs, or even data loss if the query parameter was meant to alter
the scope of a form's POST request, so it is best fixed.