Fixed #24974 -- Fixed inheritance from modelform_factory forms
When a form created using modelform_factory is inherited, the
initialization of the subclass will be different from the base class.

Consider the case when formfield_callback is passed to modelform_factory.
The formfield_callback information is lost after initialization of the
base class.

This commit throws the formfield_callback into the Meta
class of the base class, so that the subclass can use it instead
of the default.