Clarifying tutorial comment
New to django, and found this step a little confusing.
I thought it needed to be copied like the attributes above in the
question; model and template_name, however it actually means to leave
the function as it is.
Hopefully, this can make things a little clearer for newbies like me.