Fixed #25251 -- Initial data migration lost after TransactionTestCase
Data loaded in migrations are restored at the beginning of each TransactionTestCase and all the tables are truncated at the end of these test cases.
It means that, at the end of your whole test suite, if there was at least one TransactionTestCase, the migrated data are no more in the database, specially surprising when using --keepdb option.
Now we restore data at the end of a TransactionTestCase, to be sure that the next test will be in the expected environment (data loaded from initial migrations)