Fixed #26312 Added information on non-sqlite databases.
I added to the note about non-sqllite databases to suggest creating a
test_ database while creating the main database to avoid errors in
tutorial section 5 on tests.

I also added information on making sure that the user provided to the
settings file has priviledges for the two databases.  I don't have
enough experience with PostGreSQL to know whether the privileges
section applies there too, but I don't think it does.  For now it, that
part is labelled MySQL specific.