Fixed #26293 - Appending slash should return redirect in process_request
This issue was a regression that was causing appending a slash to not
return a redirect unless we also prepended a www, as the logic to
append the slash was located side the prepend WWW logic.

Moving logic required dynamically building the redirect_url as well
as updating test_non_ascii_query_string_does_not_crash to expect
a 301 from process_request instead of expecting no redirect.

Also added a regression test (and renamed another test to differentiate the two)
and release notes so this can be included in the 1.9.5 release.