Refs #16508 -- Renamed the current special-handling of "virtual" fields to "private".
The only reason why GenericForeignKey and GenericRelation are stored
separately inside _meta is that they need to be cloned for every model
subclass, but that is not true for any other virtual field. (Actually,
it's only true for GenericRelation.)

A more precise term for this concept could be "private" fields.