Include fields with col ref and raw in group by
Fix #26390.

A new (cached) property contributes_to_group_by is introduced on the
on the expression class to check whether it should be included as a
GROUP BY clause. It relies on "contains_column_references" by default,
but is explicitly set to True also for RawSQL expressions.

SQLCompiler.get_group_by is modified to use this property.