Fixed #25847 - Make User.is_(anonymous|authenticated) properties
Refactored User.is_anonymous and User.is_authenticated to make them
read-only properties instead of methods. This avoids making mistakes by
accidentally checking:

    request.user.is_authenticated

instead of:

    request.user.is_authenticated()

In order to maintain backwards compatibility, the properties' value is
a "callable boolean" as suggested by Aymeric Augustin: it behaves like
a boolean in a boolean context, but is also a callable.