Fixed #25420 -- Documented bash completion exit code behavior.
The code mentioned in #25420 is only called for bash completion. When
using bash completion the exit code isn't passed to the user, so the
value returned by Django doesn't really matter. Changed it from 1 to 0
to have it more clear that it's desired behavior and documented it.
If you're interesting in the background of this topic check #25420.