Add logging_tests for 404 warnings are logged
Simply ensuring that an existing page generates no warning, while a
non-existent page does result in a warning in the log.

Similarly, for a language-prefixed page, we also ensure that a warning
is logged if a page is not found.

All of these tests are expected to pass currently. (Notice that there
is no test for ensuring that a language-prefixed page does not
generate a warning if the page is found---there's currently a bug in
that case, so a test will be added for that along with the bug-fix in
a subsequent commit.)