Fixed #26504 -- Move logger.warning for "Not Found" until after all middlewares
This avoids logging a "Not Found" warning in the case of a request
that gets redirected to a language-prefixed URL by LocaleMiddleware.

This commit adds a test case to verify the bug fix. The test, verifies
that when a request comes in for a URL such as:

	/exists/

which LocaleMiddleware will redirect to:

	/en/exists/

That the temporary 404 result formed while middlewares are being
processed does not result in a logged warning of "Not Found:
/exists/", before LocaleMiddleware has a chance to fixup the 404 into
a 302.