Use override_settings decorator rather than "with" to reduce indentation
For any of the logging tests that need a setting of DEBUG=true
throughout an entire method, we can use an override_settings decorator
for that rather than using "with self.settings(...)". This reduces one
level of indentation for these functions.

Functions that need to this setting for only a portion of the function
body continue to use "with self.settings(...)".

Thanks to Tim Graham for recommending this change.