Refs #26445 -- Used app config to lookup user model in _create_user
Even though the UserManager is available to the user model in
migrations, one cannot use create_superuser() or create_user() due to
them using the attached model to set the password.

Since model managers used in migrations will _ALWAYS_ refer to their
current implementation and will have to work with the migration state
model and the "real" model, using the global app_registry to lookup the
user model is not a problem.