Allow users to supply their own reloader, if desired.
If the user wishes, an alternative reloader to the CPython and JYthon reloaders may be supplied in the settings.  The reloader is expected to abide by the interface of main/args/kwargs, but is otherwise free to implement whatever alternative means of connecting the parent and child processes the developer wishes.

To specify a new loader, write one in a separate file and put into the same directory as settings.py.

from my_reloader import my_reloader_function
RELOADER = { "reloader": my_reloader_function }

A dictionary, rather than a reference to the function, was chosen to give the user additional space to specify parameters that should be passed to the reloader.

This patch is provided mainly to support restarting the Django process even if it's implemented under Hy or some other syntax that generates a Python AST.