Fixed #26692 -- Relaxed a i18n compilemessages test.
Combination of:

* Peculiarities of GNU gettext tools handling of error reporting (via
  sdterr)
* The fact that for this test we need to set the `LANG` env var to `"C"`
  to ensure such error messages are emitted in English
* Special characteristics of the Windows console regarding handling of
  encoding

Makes it hard to portably test for the actual content of the `msgfmt(1)`
error message that includes the culprit Unicode code point, as exercised
by the `i18n.test_compilation.CompilationErrorHandling.test_msgfmt_error_including_non_ascii`
test.

Proposed solution is to simply drop it but still check for other parts
of the expected error message.