Fix action select color in admin changelist
The CSS defines the `background: none` for this particular select box
which makes it use white background, like the container behind it.
But it does not set any `color` property. So it uses the system default
for form controls.

Thus, when using light-on-dark system theme, it renders white text on
white background, making the text in the select box invisible or
unreadable at best.