Control visibility of add/change/delete_related admin buttons in dbfield
In certain cases, displaying the add/change/delete icons on related fields in the Django admin is undesirable, but this behavior can't currently be overridden. This patch allows configuration of which buttons to display by passing kwargs into formfield_for_db. If you are OK with the approach, I can write tests and add docs. Let me know if I should pursue. Thanks!