Fixed #26784 -- pass remote model class to router rather than the class of possibly-None model_instance
The model_instance argument to model fields' clean() and validate() methods is allowed to be None, and when this was the case, ForeignKey.validate() was passing NoneType as the first argument to router.db_for_read(). But that argument is expected to be a model class, not NoneType, so any routers that tried to access model._meta would throw an AttributeError.

Really, it makes more sense to pass in the remote model class because that's the model that's being looked up, and, conveniently, is never None.