Fixed #26761 Added tooltip for custom field in admin changelist view
A help_text property can be set for a custom field defined
in ModelAdmin.list_display. This property is displayed as a
tooltip on hover in the change list page of the admin.

Made changes in order to build Flake8 and docs.

Made changes asked in review by profuel:

- Restore space in class_attrib value, yield in
admin_list result_headers method.

- Shift space in the conditionnal in template change_list_result.