Fixed #25461: Handling class inheritance for meta API backward code examples
In case of inherited classes, get_all_related_objects and get_all_related_objects_with_model
were not returning the OneToOneField auto-created by Django to handle this inheritance.
But the examples given to help people to go through this transition smoothly are not
equivalent for these 2 functions.
We are also testing that the returned fields are not concrete to avoid this difference.