Fix dumpdata.py out of memory with queryset.iterator()
Django loads all objects into memory when iterating over a queryset, even with .iterator().
Thus, it runs out of memory when queryset.iterator() operates on all objects of a huge table (it crashes for a table with 5 million rows for me).
So I borrowed the snippet https://djangosnippets.org/snippets/1170/ written by jkocherhans, which reads a chunk of the queryset every time. It requires the queryset to be ordered, which has already been done in dumpdata.py.