Save empty gis geometries to the database.
Previously Falsey geometry objects didn't get saved, which is a random mix
depending on their types & content. Every geometry object is now explicitly
saved regardless of Truthiness.

EMPTY handling isn't that great with the supporting libraries/DBs, but this at
least always ensures the geometry is passed to the database to decide.

* Skips issue with GEOS 3.5.0 for 'POINT EMPTY'
* Skips issue with PostGIS 2.2.2 for 'POLYGON EMPTY'

Refs #26789