Fixed #26947 -- Added an option for the HSTS header preload directive
Users can now set the new `SECURE_HSTS_PRELOAD` preference to `True`, to
append the `preload` directive to  the `Strict-Transport-Security`
header. This directive is one of the pre-requisites for site owners
submitting their domain to the HSTS preload list, which helps protect
the first connection to a site, prior to the HSTS header being received.

For more information, see:
https://developer.mozilla.org/en-US/docs/Web/Security/HTTP_strict_transport_security#Preloading_Strict_Transport_Security