Fixed #27007. Handle non-UTF-8 bytes objects for text/* attachments
The fallback logic which allows non-UTF-8 encoded files to be passed to
`attach_file()` even when a `text/*` mime type has been specified has
been moved to `attach()`. Both functions now fall back to a content type
of `application/octet-stream`.

A side effect is that a file's content is decoded in memory instead of
opening it in text mode and reading it into a string.

Some mimetype-related logic in `_create_attachment()` has become
obsolete as the code moved from `attach_file()` to `attach()` already
handles this.