Fixed creation of ArrayField(RangeField) instances
This patch integrates and updates the contributions from edrmp and
dvarrazzo from Django ticket #24726. The tests pass, but I'm not sure I
put the psycopg2 adapter bits in the right place. I guess this is more a
proof of concept.

I think there needs to be more discussion about how arrays of range
fields should work. Lookups don't work right now, and I'm not sure how
they would.

Also, dvarrazzo's suggestion to subclass NumericRange works, but
wouldn't that require range objects destined for the database to be
created with the adapted subclasses? I'm not sure if that's desired for
Django. I'm interested to hear what the core team thinks.

Anyway, I hope this patch is helpful!