Fixed #27054 -- makemigrations fails with external databases
Change applied_migrations in MigrationRecorder to return an empty set
if the django_migrations table is missing and it cannot be created.

This is necessary when there are multiple database connections and at
least one of the connections is a read-only connection to an external
database that will never have migrations applied to it. All models for
the read-only database will have managed = False.