Added environment variable support to tox configuration
PYTHONPATH is required for passing through test modules that don't
already exist within the Django tests/ directory.

DISPLAY is required for testing selenium on systems that use Xvfb.

HOME is where the Vagrant djangobox stores geoip data, but it could be
useful for other tox commands that require access to data outside of
Django.