Made FieldError messages uniform across Python versions
The exception message used `repr` to format the field name. The field
name is unicode and its `repr` changes depending on the version of
Python:

- Python 2: `u'field_name'`
- Python 3: `'field_name'`

This is not handy to test and lead to the use of
`six.assertRaisesRegex()` to tolerate the presence of the leading `u`
character. A more consistent approach is to use `'%s'` in the exception
message instead of `%r`, to make sure the message is the same on both
versions.