Optimization for get_search_results() in admin
With the previous implementation of get_search_results(), Django ORM was introducing unnecessary JOINs in the final query when searching in fields of reverse relations (related issues: https://code.djangoproject.com/ticket/16603 and https://code.djangoproject.com/ticket/25789). The more words were in the search term the more inefficient was the query. This implementation seems that it makes Django ORM take a different path that does not introduce unnecessary JOINs. Searching with a great number of words in the searching term is not a problem anymore.