Removed bad name 'backend' django.db's __all__
This hasn't existed since 051c666acac770dea1af2fc223cc695f985f02f7 and thus `from django.db import *` has been broken since. This also comes with the discovery that the `pyflakes` checker for this (`flake8` code F821) doesn't work unless `__all__` is a `tuple`, presumably because if it's a `list` it's mutable.