Fixed #27334 -- TemporaryUploadedFile moved to destination when possible
When a FileField is set to an instance of File that is not also an
instance of FieldFile, pre_save will pass that object as the contents to
Storage.save().

As a result, when an uploaded file is written to temporary storage,
if possible it will be moved (rather than copied) to the upload
destination.