Fixed occasional missing plural forms in JS
This bug occurs when you have a missing translation in a language with more than two plural forms. In this case, you have a fallback to English and depending on the order in which the catalog is traversed in the for-loop you might end up with an msgid of a fallback English two-plural translation. When that happens, all translations will be reduced to two plural forms.