Added multipledispatch API to CombinedExpressions.
This allows us to overload operators depending on output_field of
operands. See CharField for an example.

Note that registration mechanism is *not* backend-specific, so
registered expression classes need to handle that on their own, using
as_<backend> as usual. This is a good thing, since we want to move away
from using backends to generate SQL.