Fixed #27775 -- Respect custom expiry in signed cookies.
The custom expiry value `_session_expiry` is stored in the session
itself. The signed cookie needs to first be loaded, and then the value
needs to be checked for validity.

This PR calls `signing.loads()` twice which may incur a small performance
penalty. The proper solution would be to refactor the session expiry
as discussed in #19201. However, this fix should not really make that
refactor any harder an may be an acceptable fix in the meantime.

Refs #19200, #19201.