Ticket #22673 -- Allow postgres schemas
Add a feature for schemas, and an sql operation for schema SQL.
Modify base code to split table names that contain a '.' into schema.table.
For every table created, create the schema if it hasn't already been created.
Enable feature for postgres, fix quoting ("schema"."table"), and fix work-around for #7783.
Added simple test for creating and quering a table in a non-standard schema.
Added corresponding documentation.