Add 'batches' function to functional utils
Refs #28231.
In effort of removing the list conversion of 'objs' in 'bulk_create'.
Added a utility function to split any iterable to multiple generators,
each of up to a specificed amount of objects in it.
Indevidual batches can then be convereted to a list independently,
hopefullying reducing memory requirement and preprocessing.

Signed-off-by: Nir Izraeli <nirizr@gmail.com>