Fixed #28231: bulk_create: no conversion of objs to list
Previous implementation of QuerySet.bulk_create forcibly converted the
iterable of objects to create to a list, regardless of the batch_size
parameter. Instead, I used newly created utils.functional.batches
function to split objects into multiple batches with a max size of
batch_size before converting each batch individually.

This does have the caveat of splitting to batches before partitioning
according to pk availability, so it has the chance of increasing the
number of actually performed queries, each having less than batch_size
objects. If desired, this may be mitigated by testing the type of objs
or providing a default-off flag. In that case, more documentation is
required.

Add myself to AUTHORS

Signed-off-by: Nir Izraeli <nirizr@gmail.com>