Changed bulk_batch_size return value definition
bulk_batch_size now returns None instead of len(objs) to mark unlimited
max batch size, but is forced to never return 0. Callers no longer need
to wrap it with a call to max(), and may instead or it with len(objs).
This is to avoid calling len(objs) on generators in bulk_create

Signed-off-by: Nir Izraeli <nirizr@gmail.com>