Make invalid AutoField/IntegerField/FloatField values easier to debug
When a large model is updated and saved with invalid values,
it produces a traceback deep within the ORM, with no clue
which field assignment caused the error. Developers are faced with:
"TypeError: int() argument must be a string, a bytes-like object or a number, not 'tuple'"
This change displays the field name which makes spotting errors a lot easier.

Trac issue 28393