Fixed #28337 -- assertRedirects correctly replays requests
Refactored django.test.clients.py:Client so that 'extra' parameters
passed into requests are stored so they can be accessed in the future.
This because useful in assertRedirects when the original request has
follow=False and the call to assertRedirects has
fetch_direct_response=True.

The problem with reported in the ticket is that under certain conditions
assertRedirect will replay the request with different parameters than
originally specified. The conditions necessary for this to occur are
passing follow=False to the original client.get and also passing 'extra'
parameters that will be interpreted as HTTP header's. Assert redirect
makes an assumption that it is possible to exactly recreate the original
request with the information that is available.

To properly recreate the the original request it would be necessary to
pass ALL method parameters AND all of the kwarg parameters that were
originally passed in. This would require storing the 'extras' in the
Client (django.test.client.py) so they could be acccessed in
assertRedrrects and properly replayed.

I believe this is the proper solution because it removes the assumptions
present when assertRedirect attempts to recreate the request. Although
this is a corner case, it is certainly acceptable for django users to
utilize the infinite space that is HTTP headers and provide varrying
logic depending on the values present. The downside of this
implementation is that it bloats the Client by requiring it to store the
extra parameters passed in. These extra parameters will rarely be used
but I believe the tradeoff of adding properties to Client and
potentially increasing memory footprint are small enough to warrant this
change.