Refs #28010 -- Allowed reverse related fields in select_for_update(of=(...)).
Previously, reverse related fields (OneToOneFields in this case) were
not matched correctly and thus effectively excluded, even though they
are allowed in select_related().

Reverse OneToOne fields are normally followed using a LEFT JOIN, which
are not allowed in SELECT FOR UPDATE. Thus, in order to use this
support, it is necessary to force an INNER JOIN, as is demonstrated in
the test.

Thanks to Adam Chidlow for the report and suggested fix.