admin: changeform view: handle ValidationErrors
This patch handles ValidationErrors coming from `save_model` or
`save_related`.

This is useful for e.g. django-fsm, where a transition method might
throw this.
An alternative approach would be to hook into the form's validation
(`is_valid`), but this is not as straight forward and it seems like a
good behavior to handle ValidationErrors here, instead of letting them
bubble up and typically causing a 500 for the admin page.

Ref: https://github.com/kmmbvnr/django-fsm/issues/120