Make the construction of the endpoint regex pattern stable.
It is possible to query the endpoints that your Django app provides.

    from django.core import urlresolvers
    urlresolvers.get_resolver()

One can then walk that tree to enumerate each of the regex patterns that went into constructing it.

The regex pattern that is constructed for an endpoint is not stable. For instance, in my project I have seen it construct each of the following:

^admin/^(?P<app_label>social_django|authtoken|queues|jobs|files|users|clusters|appps|auth|vdcs)/$
^admin/^(?P<app_label>social_django|auth|users|vdcs|appps|jobs|queues|authtoken|clusters|files)/$
^admin/^(?P<app_label>authtoken|appps|social_django|clusters|users|queues|jobs|vdcs|auth|files)/$
^admin/^(?P<app_label>authtoken|appps|social_django|clusters|users|queues|vdcs|jobs|auth|files)/$

This precludes the use of patterns in unit tests.

Make the construction of the patterns stable so that subsequent walks of the tree will be consistent.