Fixed #28712 -- Add option attributes to ChoiceWidget.
The django.forms.ChoiceWidget and all its subclasses such as
django.forms.Select now optionally accept a parameter `opt_attrs`
which applies a dictionary of attributes to the `option` HTML nodes
generated by the widget. This is similar to how the existing `attrs`
parameter works. Developers can use this to style options if they so
desire.