[WIP] Keep ModelAdmin's JS/CSS before form's
The changes in https://code.djangoproject.com/ticket/28377 might cause
unexpected ordering with django-autocomplete-light on an admin page that
also includes a date selector widget.

The reason for this is that the date widget includes
`'admin/js/vendor/jquery/jquery.js'`, which is also included in the
Admin's media already, and then everything before it gets prepended.

Simplified version:

    admin_media = ['jquery.js']
    form_media = ['app/jquery.init.js', 'jquery.js', 'calendar.js']
    merged = ['app/jquery.init.js', 'jquery.js', 'calendar.js']

This causes a JavaScript error, since `app/jquery.init.js` has no jQuery
available.

A solution to this might be to keep the admin's media separate from the
form's media, which this PR tries to do.