Add tests for custom user forms regression
As described in [#29449](https://code.djangoproject.com/ticket/29449)
those are the tests that should still pass with the recent change in
3333d935d2914c.

The regression is about assuming CharField subclass for
username field on a form. Those tests use the already existing
`CustomUserWithFK` testing model that has username field as a
foreign key to the different `Email` model.