Preparation for Filter Expressions/Expression Filters.
This is in preparation for https://github.com/django/django/pull/8119.

In order to make it possible to have different behaviour when an
expression is to be used _only_ in a WHERE clause, we need to have
some tag on the expression.

I think maybe we need to make the resolve_expression() accept arbitrary
kwargs - it's either that or just add for_where=False.