Fixed #11927 -- Add dumpdata --yaml-collection-style option
By default, pyyaml uses the "block" style (separate line per element)
for collections that contain nested collections, and the "flow" style
(single line for all elements) for other collections. However, when
creating fixtures or otherwise hand-editing or versioning the result of
dumpdata, the block style is preferred.

This commit adds a new option for the `dumpdata` command which allows
forcing the use of either the 'block' or 'flow' style. The default is
'mixed' and thus unchanged.