Refs #11964 -- Changed CheckConstraint() signature to use keyword-only arguments.
Using keyword only arguments follows the `Index.__init__` pattern signature to
easen the addition of constraint options in the future.

Also renamed the `constraint` argument to `check` to better represent which
part of the constraint the provided `Q` object represents.

This change will be particularily useful when we add partial constraint support
because an additional `Q` object passing be required.